feat(AuthController, UsersController, SecurityConfig): Add authentication class and fix routes in SecurityConfig
