+++ src/main/java/com/QuickNotesAPI/QuickNotes/controller/AuthController.java
@@ -1,0 +1,45 @@
+package com.QuickNotesAPI.QuickNotes.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.QuickNotesAPI.QuickNotes.model.Users;
+import com.QuickNotesAPI.QuickNotes.service.UsersService;
+
+@RestController
+@RequestMapping("/api/auth")
+public class AuthController {
+    @Autowired
+    private UsersService usersService;
+
+    @PostMapping("/register")
+    private ResponseEntity<String> register(@RequestBody Users user) {
+        try {
+            usersService.register(user);
+            return ResponseEntity.status(HttpStatus.CREATED).body("User registered successfully");
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("An unexpected error occurred: " + e.getMessage());
+        }
+    }
+
+    @PostMapping("/login")
+    private ResponseEntity<String> login(@RequestBody Users user) {
+        try {
+            String email = user.getEmail();
+            String password = user.getUsersPassword();
+            String jwt = usersService.login(email, password);
+            return ResponseEntity.ok(jwt);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("An unexpected error occurred: " + e.getMessage());
+        }
+    }
+}
